add multi role

+++ routes/web.php
@@ -3,6 +3,7 @@
 use App\Http\Controllers\AuthController;
 use App\Http\Controllers\EmployeeController;
 use App\Http\Controllers\JobController;
+use App\Http\Controllers\RoleController;
 use Illuminate\Support\Facades\Route;
 
 /*
@@ -15,23 +16,26 @@ use Illuminate\Support\Facades\Route;
 | contains the "web" middleware group. Now create something great!
 |
 */
+
 Route::get('/login', [AuthController::class,'login'])->name('login')->middleware('guest');
 Route::post('/login', [AuthController::class,'auth']);
 Route::get('/logout', [AuthController::class,'logout'])->middleware('auth');
 
 Route::controller(EmployeeController::class)->group(function () {
     Route::get('/', 'index')->middleware('auth');
-    Route::get('/employee/{id}', 'show')->middleware('auth');
+    Route::get('/employee/{id}', 'show')->middleware(['auth', 'must-admin']);
 
-    Route::get('/employee-add', 'create')->middleware('auth');
-    Route::post('/employee', 'store')->middleware('auth');
+    Route::get('/employee-add', 'create')->middleware(['auth', 'must-admin']);
+    Route::post('/employee', 'store')->middleware(['auth', 'must-admin']);
 
-    Route::get('/employee-edit/{id}', 'edit')->middleware('auth');
-    Route::put('/employee/{id}', 'update')->middleware('auth');
+    Route::get('/employee-edit/{id}', 'edit')->middleware(['auth', 'must-admin']);
+    Route::put('/employee/{id}', 'update')->middleware(['auth', 'must-admin']);
 
-    Route::put('/employee/{id}', 'update')->middleware('auth');
+    Route::put('/employee/{id}', 'update')->middleware(['auth', 'must-admin']);
 
-    Route::delete('/employee-destroy/{id}', 'destroy')->middleware('auth');
+    Route::delete('/employee-destroy/{id}', 'destroy')->middleware(['auth', 'must-admin']);
 });
 
-Route::get('/jobs', [JobController::class,'index'])->middleware('auth');
+Route::get('/jobs', [JobController::class,'index'])->middleware(['auth', 'must-admin']);
+
+Route::get('/roles', [RoleController::class,'index'])->middleware(['auth', 'must-admin']);
